+++ lib/rmodbus/common.rb
@@ -15,11 +15,9 @@
 
 module ModBus
   module Common
-    # @return [Boolean] debug mode
-    # default false
-    attr_accessor :debug
+    attr_accessor :debug, :raise_exception_on_mismatch, 
+                  :read_retries, :read_retry_timeout 
 
-    @debug = false
 
     private
     # Put log message on standart output